add support for Limbo.
it adds a parser, adjusts the detector (there is now extension clash
between matlab/limbo/objective c, and another for for limbo/basic).
there are also sample limbo files (in public domain) and a test.
all tests pass with this.

it has been tested on a large body of limbo code (mixed with C,
makefile, shell, awk):

	http://code.google.com/p/inferno-os/

more info on Limbo:

	http://www.vitanuova.com/inferno/papers/limbo.html