[NEW] Gestalts for silverlight + popular JS libs
New gestalts for:
	silverlight
	mootools
	prototype
	yahoo ui (yui)
	dojo

Implementing the silverlight gestalt requires parsing *.csproj files.

There appears to be a bug in the ohcount C core. If you query
the language_breakdowns of a source_file object, but that source_file
happens to not contain any known languages, then the source_file.contents
are reset to nil! I gave this bug a good long hunt, but I just couldn't
find it.

To work around this bug, *.csproj files are now counted as XML. It's a
bit cheeky, but they really are XML, so it works out. Without this,
the gestalt rule would explode because the contents are nil.