[FIX] Ruby binding memory performance improvement
This patch corrects the terrible memory thrashing that occurs when
Ohcount is called through the Ruby binding layer.

Each call into Ohcount is accompanied by an array of filenames. This
list of filenames is used for hinting in detector.c.

For each call into Ohcount, The Ruby binding layer allocated and then
freed a clone of this array of filenames two times -- once in the SWIG
layer, and once in sourcefile.c.

This is OK for small arrays, but in some cases we have thousands of
filenames. This caused a massive amount of heap churn and, eventually,
resulted in kernel OOM panics.

Ohcount no longer clones the list of filenames. Instead, we just use
pointers into the original Ruby strings. Presto: no more heap churn; a
huge boost in speed; and no need to force a Ruby GC after every call
into the library.