[FIX] Emacs mode "C" improperly detected as C++
detector.c uses the file extenstion map to convert emacs modes
to languages. Because the upper-case "C" file extension maps to
C++, this means that emacs mode "C" also mapped to C++, which was
incorrect.

I've added a special check in the code for this case. Emacs modes
"c" and "C" now correctly map to straight C code.