Update Ohcount to work with Ubuntu 14
The most significant change is that the perl script
that generates HTML, especially with the "<!DOCTYPE html" heading,
is now identified as HTML.  This is due to differences in the
libmagic database between CentOS 5 and Ubuntu 14.

It may be possible to generate a different database, or modify the
existing database to change the identification order.  However,
it is also worth acknowledging that, in the years since this code
was first developed, it may be more correct to recognize a script
that generates a complete HTML document as 'HTML', even if that
script is written in something like perl.