OTWO-3061 Fix mathematica comment parsing
The original code in PR#23 did not consider multiline comments. We
compared and copied the code from OCaml parser which has a similar
comment structure and it worked.
The existing test files in test/expected_dir & test/src_dir needed
fixing for some mismatch in whitespace characters and expected lines.