OTWO-3061 Octave comments nested within mathematica
When octave comments were nested within mathematica comments, the parser
incorrectly detected the file as Octave. To fix this we moved around an
If/Else block which placed mathematica at the bottom of the conditional.
We have also tried adding mathematica comments within Octave comments to
see if the reverse was true. We found that Octave parsing worked
correctly without any further change in code.